Remove ForbiddenCharacters to Converters
